sb/intel: Fix pointer casts
Fix some compiler warnings due to pointer to integer conversions
with different size.
Required for 64bit ramstage.

Reviewed-on: https://review.coreboot.org/c/29875
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/southbridge/intel/common/smi.c
@@ -150,7 +150,7 @@ void smm_setup_structures(void *gnvs, void *tcg, void *smi1)
 		"outb %%al, %%dx\n\t"
 		: /* ignore result */
 		: "a" (APM_CNT_GNVS_UPDATE),
-		  "b" ((u32)gnvs),
+		  "b" ((uintptr_t)gnvs),
 		  "d" (APM_CNT)
 	);
 }
