+++ src/southbridge/intel/i82801ix/lpc.c
@@ -553,7 +553,7 @@ static void southbridge_inject_dsdt(struct device *dev)
 
 		/* Add it to SSDT.  */
 		acpigen_write_scope("\\");
-		acpigen_write_name_dword("NVSA", (u32) gnvs);
+		acpigen_write_name_dword("NVSA", (uintptr_t)gnvs);
 		acpigen_pop_len();
 	}
 }
